requirements.txt.txt

python-telegram-bot[job-queue]==20.3
requests
aiohttp